requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
ta-lib>=0.4.0
backtrader>=1.9.0
jupyter>=1.0.0
statsmodels>=0.14.0
scipy>=1.11.0
ccxt>=4.0.0
plotly>=5.17.0
feedparser>=6.0.10
vaderSentiment>=3.3.2
python-dotenv>=1.0.0
tqdm>=4.65.0
xgboost>=2.0.0
joblib>=1.3.0